requirements.txt
fastapi
uvicorn
httpx
gspread
google-auth
apscheduler
python-dotenv